+++ app/src/main/java/co/kr/woowahan_repo/domain/repository/GithubIssuesRepository.kt
@@ -1,4 +1,6 @@
-package co.kr.woowahan_repo.domain
+package co.kr.woowahan_repo.domain.repository
+
+import co.kr.woowahan_repo.domain.entity.GithubIssueModel
 
 interface GithubIssuesRepository {
     suspend fun fetchIssues(state: String, page: Int): List<GithubIssueModel>
