+++ app/src/main/java/co/kr/woowahan_repo/presentation/viewmodel/IssuesViewModel.kt
@@ -1,8 +1,8 @@
 package co.kr.woowahan_repo.presentation.viewmodel
 
 import androidx.lifecycle.*
-import co.kr.woowahan_repo.domain.GithubIssuesRepository
-import co.kr.woowahan_repo.domain.GithubIssueModel
+import co.kr.woowahan_repo.domain.repository.GithubIssuesRepository
+import co.kr.woowahan_repo.domain.entity.GithubIssueModel
 import kotlinx.coroutines.launch
 import timber.log.Timber
